Gemma3/requirements.txt
openai-whisper
requests
sounddevice
numpy
faiss-cpu
sentence-transformers
portaudio
pyaudio 